requirements.txt
# Updated requirements.txt for backend

annotated-types==0.7.0
blinker==1.9.0
cachetools==5.5.2
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.2.1
Flask==3.1.1
flask-cors==6.0.0
Flask-SQLAlchemy==3.1.1
google-ai-generativelanguage==0.6.15
google-api-core==2.25.1
google-api-python-client==2.172.0
google-auth==2.40.3
google-auth-httplib2==0.2.0
google-generativeai==0.8.5
googleapis-common-protos==1.70.0
greenlet==3.2.3
grpcio==1.73.0
grpcio-status==1.71.0
httplib2==0.22.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
pillow==11.2.1
proto-plus==1.26.1
protobuf==5.29.5
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.11.7
pydantic_core==2.33.2
pyparsing==3.2.3
requests==2.32.4
rsa==4.9.1
SQLAlchemy==2.0.41
tqdm==4.67.1
typing-inspection==0.4.1
typing_extensions==4.14.0
uritemplate==4.2.0
urllib3==2.4.0
Werkzeug==3.1.3

# New dependencies for authentication and community features
firebase-admin==6.6.0
python-jose==3.3.0
bcrypt==4.2.0
Flask-JWT-Extended==4.7.1
marshmallow==3.24.0
marshmallow-sqlalchemy==1.2.0
google-cloud-firestore==2.21.0
google-cloud-storage==2.19.0
redis==5.3.0
celery==5.4.0
Flask-Mail==0.10.0
python-dotenv==1.0.1